move hooks to core library
