+++ core/lib/hooks.ts
